+++ src/store/reducers/app-process/selectors.jsx
@@ -1,0 +1,15 @@
+export const getSelectedCity = (state) => {
+  return state.PROCESS.selectedCity;
+};
+
+export const getActiveOfferId = (state) => {
+  return state.PROCESS.activeOfferId;
+};
+
+export const getSortType = (state) => {
+  return state.PROCESS.sortType;
+};
+
+export const isOpenSortMenu = (state) => {
+  return state.PROCESS.isOpenSortMenu;
+};
