+++ src/store/reducers/user/selectors.jsx
@@ -1,0 +1,7 @@
+export const getUserLogin = (state) => {
+  return state.USER.login;
+};
+
+export const isLoggedIn = (state) => {
+  return state.USER.loggedIn;
+};
